Graphics

+++ app/src/main/java/Graphics/Quadrilateral.java
@@ -7,9 +7,9 @@ package Graphics;
 public class Quadrilateral {
     float x, y, z;
     float width, height;
-    float[] color;
+    byte[] color;
 
-    public static Quadrilateral getQuad(Quadrilateral quad, float xCoord, float yCoord, float zCoord, float w, float h, float[] c){
+    public static Quadrilateral getQuad(Quadrilateral quad, float xCoord, float yCoord, float zCoord, float w, float h, byte[] c){
         quad.x = xCoord;
         quad.y = yCoord;
         quad.z = zCoord;
